Rename source folder
now that we expect users to import individual modules, `lib` _seems_ to
be a more conventional choice
