+++ lib/ascii.js
