+++ lib/legacy.js
